pyproject.toml
[project]
name = "timelines-mcp"
version = "0.1.0"
description = "MCP server to aid LLMs in maintaining coherent long generations for time dependent narratives"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=0.2.0",
    "asyncpg>=0.29.0",
    "qdrant-client>=1.7.0",
    "chromadb>=0.4.0",
    "aiosqlite>=0.19.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-recording>=0.13.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "polyfactory>=2.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
strict = true

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 100
show_missing = true

[tool.poe.tasks]
test = "pytest --cov=src --cov-report=term-missing --cov-report=html"
format = "ruff check --fix . && ruff format ."

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "polyfactory>=3.1.0",
    "poethepoet>=0.31.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-recording>=0.13.4",
    "ruff>=0.14.8",
]
